Move DeletePolicy out of the account object
