+++ feature/migration/qrcode/src/main/kotlin/app/k9mail/feature/migration/qrcode/domain/entity/AccountData.kt
@@ -4,7 +4,7 @@ import app.k9mail.core.common.mail.EmailAddress
 import app.k9mail.core.common.net.Hostname
 import app.k9mail.core.common.net.Port
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.account.Account.DeletePolicy
+import app.k9mail.legacy.account.DeletePolicy
 
 internal data class AccountData(
     val sequenceNumber: Int,
