+++ legacy/common/src/main/java/com/fsck/k9/account/DefaultDeletePolicyProvider.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.account
 
 import app.k9mail.core.common.mail.Protocols
-import app.k9mail.legacy.account.Account.DeletePolicy
+import app.k9mail.legacy.account.DeletePolicy
 
 class DefaultDeletePolicyProvider : DeletePolicyProvider {
     override fun getDeletePolicy(accountType: String): DeletePolicy {
