+++ legacy/common/src/test/java/com/fsck/k9/account/DefaultDeletePolicyProviderTest.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.account
 
 import app.k9mail.core.common.mail.Protocols
-import app.k9mail.legacy.account.Account
+import app.k9mail.legacy.account.DeletePolicy
 import assertk.assertFailure
 import assertk.assertThat
 import assertk.assertions.isEqualTo
@@ -15,21 +15,21 @@ class DefaultDeletePolicyProviderTest {
     fun `getDeletePolicy with IMAP should return ON_DELETE`() {
         val result = deletePolicyProvider.getDeletePolicy(Protocols.IMAP)
 
-        assertThat(result).isEqualTo(Account.DeletePolicy.ON_DELETE)
+        assertThat(result).isEqualTo(DeletePolicy.ON_DELETE)
     }
 
     @Test
     fun `getDeletePolicy with POP3 should return NEVER`() {
         val result = deletePolicyProvider.getDeletePolicy(Protocols.POP3)
 
-        assertThat(result).isEqualTo(Account.DeletePolicy.NEVER)
+        assertThat(result).isEqualTo(DeletePolicy.NEVER)
     }
 
     @Test
     fun `getDeletePolicy with demo should return ON_DELETE`() {
         val result = deletePolicyProvider.getDeletePolicy("demo")
 
-        assertThat(result).isEqualTo(Account.DeletePolicy.ON_DELETE)
+        assertThat(result).isEqualTo(DeletePolicy.ON_DELETE)
     }
 
     @Test
