+++ legacy/core/src/main/java/com/fsck/k9/AccountPreferenceSerializer.kt
@@ -1,7 +1,6 @@
 package com.fsck.k9
 
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.account.Account.DeletePolicy
 import app.k9mail.legacy.account.Account.Expunge
 import app.k9mail.legacy.account.Account.MessageFormat
 import app.k9mail.legacy.account.Account.QuoteStyle
@@ -22,6 +21,7 @@ import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_STRIP
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_SYNC_INTERVAL
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.NO_OPENPGP_KEY
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.UNASSIGNED_ACCOUNT_NUMBER
+import app.k9mail.legacy.account.DeletePolicy
 import app.k9mail.legacy.account.FolderMode
 import app.k9mail.legacy.account.Identity
 import app.k9mail.legacy.notification.NotificationLight
