+++ legacy/core/src/main/java/com/fsck/k9/controller/MessagingController.java
@@ -29,7 +29,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 import app.k9mail.core.featureflag.FeatureFlagProvider;
 import app.k9mail.legacy.account.Account;
-import app.k9mail.legacy.account.Account.DeletePolicy;
+import app.k9mail.legacy.account.DeletePolicy;
 import app.k9mail.legacy.di.DI;
 import app.k9mail.legacy.message.controller.MessageReference;
 import app.k9mail.legacy.message.controller.MessagingControllerMailChecker;
