+++ legacy/core/src/main/java/com/fsck/k9/preferences/AccountSettingsDescriptions.java
@@ -12,7 +12,7 @@ import android.content.Context;
 import app.k9mail.legacy.account.AccountDefaultsProvider;
 import app.k9mail.legacy.notification.NotificationLight;
 import app.k9mail.legacy.account.Account;
-import app.k9mail.legacy.account.Account.DeletePolicy;
+import app.k9mail.legacy.account.DeletePolicy;
 import app.k9mail.legacy.account.Account.Expunge;
 import app.k9mail.legacy.account.FolderMode;
 import app.k9mail.legacy.account.Account.MessageFormat;
