+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/settings/account/AccountSettingsDataStore.kt
@@ -3,6 +3,7 @@ package com.fsck.k9.ui.settings.account
 import androidx.preference.PreferenceDataStore
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.Account.SpecialFolderSelection
+import app.k9mail.legacy.account.DeletePolicy
 import app.k9mail.legacy.notification.NotificationLight
 import app.k9mail.legacy.notification.NotificationVibration
 import com.fsck.k9.Preferences
@@ -152,7 +153,7 @@ class AccountSettingsDataStore(
                     reschedulePoll()
                 }
             }
-            "delete_policy" -> account.deletePolicy = Account.DeletePolicy.valueOf(value)
+            "delete_policy" -> account.deletePolicy = DeletePolicy.valueOf(value)
             "expunge_policy" -> account.expungePolicy = Account.Expunge.valueOf(value)
             "max_push_folders" -> account.maxPushFolders = value.toInt()
             "idle_refresh_period" -> account.idleRefreshMinutes = value.toInt()
